web_scrap_selenium/requirements.txt
2captcha-python==2.0.2
aiofiles==25.1.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
selenium==4.39.0
setuptools==80.10.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
TwoCaptcha==0.0.1
typing_extensions==4.15.0
undetected-chromedriver==3.5.5
urllib3==2.6.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.4
wsproto==1.3.2
